base-events branch cleanup
