+++ Assets/Scripts/Events/EventResponse.cs
@@ -1,0 +1,14 @@
+namespace GudKoodi.DeeperSkeeper.Event
+{
+    using System;
+    using UnityEngine;
+    using UnityEngine.Events;
+
+    /// <summary>
+    /// Mandatory Unity overhead.
+    /// </summary>
+    [Serializable]
+    public class EventResponse : UnityEvent<GameObject>
+    {
+    }
+}
