+++ Assets/Scripts/Events/EventResponse.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2c3a60892714843459a5fb9760c6a933
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
