+++ Assets/Scripts/Events/LevelGenerationRequestResponse.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 426ca2acfdb4a6e45b8e7654ff7d97b5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
