+++ Assets/Scripts/Events/ObjectCreationResponse.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7da9efab7155e484f88604bb97bf07ed
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
