+++ Assets/Scripts/Events/ObjectUpdateRequestResponse.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 05f3ffd0e6cf941498a149a720844e1e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
